chore: update reducer, saga
