+++ packages/developer-portal/src/core/store.ts
@@ -22,7 +22,7 @@ import integrationTypes from '@/reducers/app-integration-types'
 import webhookEditReducer from '../reducers/webhook-edit-modal'
 import webhookSubscriptions from '@/reducers/webhook-subscriptions'
 import developerSubscriptions from '@/reducers/developer-subscriptions'
-import membersReducer from '@/reducers/members'
+import developersReducer from '@/reducers/developers'
 
 import authSagas from '@/sagas/auth'
 import appsSaga from '@/sagas/apps/apps'
@@ -43,7 +43,7 @@ import webhookSubscriptionsSagas from '@/sagas/webhook-subscriptions'
 import { injectSwitchModeToWindow } from '@reapit/elements'
 import webhookEditSagas from '../sagas/webhook-edit-modal'
 import developerSubscriptionsSagas from '../sagas/developer-subscriptions'
-import membersSagas from '../sagas/developers/members'
+import developersSagas from '../sagas/developers'
 
 export class Store {
   static _instance: Store
@@ -82,7 +82,7 @@ export class Store {
     webhookEdit: webhookEditReducer,
     webhooks: webhookSubscriptions,
     developerSubscriptions,
-    members: membersReducer,
+    developers: developersReducer,
   })
 
   static sagas = function*() {
@@ -104,7 +104,7 @@ export class Store {
       fork(webhookEditSagas),
       fork(webhookSubscriptionsSagas),
       fork(developerSubscriptionsSagas),
-      fork(membersSagas),
+      fork(developersSagas),
     ])
   }
