+++ packages/developer-portal/src/reducers/developers/index.ts
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux'
+import membersReducer, { MembersState } from './members'
+
+export interface DevelopersRootState {
+  members: MembersState
+}
+
+export default combineReducers<DevelopersRootState>({
+  members: membersReducer,
+})
