+++ packages/developer-portal/src/sagas/developers/__tests__/index.ts
@@ -1,0 +1,11 @@
+import membersSagas from '../members'
+import { all, fork } from 'redux-saga/effects'
+import developersSagas from '../index'
+
+describe('developersSagas', () => {
+  it('should listen saga', () => {
+    const gen = developersSagas()
+    expect(gen.next().value).toEqual(all([fork(membersSagas)]))
+    expect(gen.next().done).toBe(true)
+  })
+})
