+++ packages/developer-portal/src/selector/developers.ts
@@ -1,0 +1,9 @@
+import { ReduxState } from '@/types/core'
+
+export const selectOrganisationMembers = (state: ReduxState) => {
+  return state.developers?.members?.data?.data || []
+}
+
+export const selectOrganisationMembersLoading = (state: ReduxState) => {
+  return state.developers?.members?.loading
+}
